A bit of refactoring.
